Move gcconfig.h to gc/include and install it
Moving from gc/include/private/gcconfig.h
