+++ gc/Makefile.am
@@ -180,7 +180,7 @@ CXXFLAGS += $(CFLAGS_EXTRA)
 ## FIXME: relies on internal code generated by automake.
 ## FIXME: ./configure --enable-dependency-tracking should be used
 #all_objs = @addobjs@ $(libgc_la_OBJECTS)
-#$(all_objs) : include/private/gcconfig.h include/private/gc_priv.h \
+#$(all_objs) : include/gcconfig.h include/private/gc_priv.h \
 #include/private/gc_hdrs.h include/gc.h include/gc_gcj.h \
 #include/gc_pthread_redirects.h include/gc_config_macros.h \
 #include/private/thread_local_alloc.h include/private_support.h \
