+++ gc/doc/README.macros
@@ -20,7 +20,7 @@ predefined macros.  Currently these macros generally identify platforms
 instead of features.  In many cases, this is a mistake.
 
 Many of the tested configuration macros are at least somewhat defined in
-either include/private/gcconfig.h or in Makefile.direct.  Here is an attempt
+either include/gcconfig.h or in Makefile.direct.  Here is an attempt
 at documenting these macros:  (Thanks to Walter Bright for suggesting
 this.)
