+++ gc/include/include.am
@@ -15,6 +15,7 @@
 pkginclude_HEADERS += \
         include/gc.h \
         include/gc_backptr.h \
+        include/gcconfig.h \
         include/gc_config_macros.h \
         include/gc_inline.h \
         include/gc_mark.h \
@@ -36,7 +37,6 @@ dist_noinst_HEADERS += \
         include/private/gc_locks.h \
         include/private/gc_pmark.h \
         include/private/gc_priv.h \
-        include/private/gcconfig.h \
         include/private/pthread_stop_world.h \
         include/private/pthread_support.h \
         include/private/specific.h \
